rename usePagedBundleAssets just to useBundleAssets
