+++ src/services/bundleAnalysis/useBundleAssets.tsx
@@ -1,16 +1,20 @@
-import { useQuery } from '@tanstack/react-query'
-import isNull from 'lodash/isNull'
+import { useInfiniteQuery } from '@tanstack/react-query'
 import { z } from 'zod'
 
 import { MissingHeadReportSchema } from 'services/comparison'
 import {
   RepoNotFoundErrorSchema,
   RepoOwnerNotActivatedErrorSchema,
-  useRepoOverview,
 } from 'services/repo'
 import Api from 'shared/api'
 import { type NetworkErrorObject } from 'shared/api/helpers'
+import { mapEdges } from 'shared/utils/graphql'
 import A from 'ui/A'
+
+const PageInfoSchema = z.object({
+  hasNextPage: z.boolean(),
+  endCursor: z.string().nullable(),
+})
 
 const BundleDataSchema = z.object({
   loadTime: z.object({
@@ -48,7 +52,10 @@ const BundleAssetSchema = z.object({
   measurements: AssetMeasurementsSchema.nullable(),
 })
 
-type BundleAsset = z.infer<typeof BundleAssetSchema>
+const BundleAssetPaginatedSchema = z.object({
+  edges: z.array(z.object({ node: BundleAssetSchema })),
+  pageInfo: PageInfoSchema,
+})
 
 const BundleAnalysisReportSchema = z.object({
   __typename: z.literal('BundleAnalysisReport'),
@@ -59,7 +66,7 @@ const BundleAnalysisReportSchema = z.object({
           uncompress: z.number(),
         }),
       }),
-      assets: z.array(BundleAssetSchema),
+      assetsPaginated: BundleAssetPaginatedSchema.nullable(),
     })
     .nullable(),
 })
@@ -103,9 +110,12 @@ query BundleAssets(
   $branch: String!
   $bundle: String!
   $interval: MeasurementInterval!
-  $before: DateTime!
-  $after: DateTime
+  $dateBefore: DateTime!
+  $dateAfter: DateTime
   $filters: BundleAnalysisReportFilters
+  $assetsAfter: String
+  $orderingDirection: OrderingDirection
+  $ordering: AssetOrdering
 ) {
   owner(username: $owner) {
     repository(name: $repo) {
@@ -122,34 +132,47 @@ query BundleAssets(
                       uncompress
                     }
                   }
-                  assets {
-                    name
-                    extension
-                    bundleData {
-                      loadTime {
-                        threeG
-                        highSpeed
-                      }
-                      size {
-                        uncompress
-                        gzip
-                      }
-                    }
-                    measurements(
-                      interval: $interval
-                      before: $before
-                      after: $after
-                      branch: $branch
-                    ) {
-                      change {
-                        size {
-                          uncompress
+                  assetsPaginated(
+                    first: 20
+                    after: $assetsAfter
+                    orderingDirection: $orderingDirection
+                    ordering: $ordering
+                  ) {
+                    edges {
+                      node {
+                        name
+                        extension
+                        bundleData {
+                          loadTime {
+                            threeG
+                            highSpeed
+                          }
+                          size {
+                            uncompress
+                            gzip
+                          }
+                        }
+                        measurements(
+                          interval: $interval
+                          before: $dateBefore
+                          after: $dateAfter
+                          branch: $branch
+                        ) {
+                          change {
+                            size {
+                              uncompress
+                            }
+                          }
+                          measurements {
+                            timestamp
+                            avg
+                          }
                         }
                       }
-                      measurements {
-                        timestamp
-                        avg
-                      }
+                    }
+                    pageInfo {
+                      hasNextPage
+                      endCursor
                     }
                   }
                 }
@@ -175,15 +198,17 @@ interface UseBundleAssetsArgs {
   provider: string
   owner: string
   repo: string
-  branch?: string
+  branch: string
   bundle: string
   interval?: 'INTERVAL_1_DAY' | 'INTERVAL_7_DAY' | 'INTERVAL_30_DAY'
-  before?: Date
-  after?: Date | null
+  dateBefore?: Date
+  dateAfter?: Date | null
   filters?: {
     reportGroups?: string[]
     loadTypes?: string[]
   }
+  orderingDirection?: 'ASC' | 'DESC'
+  ordering?: 'NAME' | 'SIZE' | 'TYPE'
   opts?: {
     enabled?: boolean
     suspense?: boolean
@@ -194,35 +219,17 @@ export const useBundleAssets = ({
   provider,
   owner,
   repo,
-  branch: branchArg,
+  branch,
   bundle,
   interval,
-  before,
-  after,
+  dateBefore,
+  dateAfter,
   filters = {},
+  orderingDirection,
+  ordering,
   opts,
 }: UseBundleAssetsArgs) => {
-  const { data: repoOverview, isSuccess } = useRepoOverview({
-    provider,
-    repo,
-    owner,
-    opts: {
-      enabled: !branchArg,
-    },
-  })
-
-  let enabled = true
-  if (opts?.enabled !== undefined) {
-    enabled = opts.enabled
-  }
-
-  if (opts?.enabled !== undefined && !branchArg) {
-    enabled = opts.enabled && isSuccess
-  }
-
-  const branch = branchArg ?? repoOverview?.defaultBranch
-
-  return useQuery({
+  const { data, ...rest } = useInfiniteQuery({
     queryKey: [
       'BundleAssets',
       provider,
@@ -231,24 +238,29 @@ export const useBundleAssets = ({
       branch,
       bundle,
       interval,
-      before,
-      after,
+      dateBefore,
+      dateAfter,
       filters,
+      ordering,
+      orderingDirection,
     ],
-    queryFn: ({ signal }) =>
+    queryFn: ({ signal, pageParam }) =>
       Api.graphql({
+        query,
         provider,
         signal,
-        query,
         variables: {
           owner,
           repo,
           branch,
           bundle,
           interval,
-          before,
-          after,
+          dateBefore,
+          dateAfter,
           filters,
+          assetsAfter: pageParam,
+          ordering,
+          orderingDirection,
         },
       }).then((res) => {
         const parsedData = RequestSchema.safeParse(res?.data)
@@ -257,7 +269,7 @@ export const useBundleAssets = ({
           return Promise.reject({
             status: 404,
             data: {},
-            dev: 'useBundleAssets - 404 Failed to parse schema',
+            dev: 'useBundleAssets - 404 schema parsing failed',
           } satisfies NetworkErrorObject)
         }
 
@@ -267,8 +279,7 @@ export const useBundleAssets = ({
           return Promise.reject({
             status: 404,
             data: {},
-            dev: 'useBundleAssets - 404 Not found error',
-          } satisfies NetworkErrorObject)
+          })
         }
 
         if (data?.owner?.repository?.__typename === 'OwnerNotActivatedError') {
@@ -284,27 +295,48 @@ export const useBundleAssets = ({
                 </p>
               ),
             },
-            dev: 'useBundleAssets - 403 Owner not activated',
-          } satisfies NetworkErrorObject)
+          })
         }
 
-        let assets: Array<BundleAsset> = []
-        let bundleUncompressSize: number | null = null
         if (
+          !data?.owner ||
           data?.owner?.repository?.branch?.head?.bundleAnalysisReport
-            ?.__typename === 'BundleAnalysisReport' &&
-          !isNull(data.owner.repository.branch.head.bundleAnalysisReport.bundle)
+            ?.__typename === 'MissingHeadReport'
         ) {
-          bundleUncompressSize =
-            data.owner.repository.branch.head.bundleAnalysisReport.bundle
-              .bundleData.size.uncompress
-          assets =
-            data.owner.repository.branch.head.bundleAnalysisReport.bundle.assets
+          return {
+            assets: [],
+            bundleData: null,
+            pageInfo: null,
+          }
         }
 
-        return { assets, bundleUncompressSize }
+        const assets = mapEdges(
+          data?.owner?.repository?.branch?.head?.bundleAnalysisReport?.bundle
+            ?.assetsPaginated
+        )
+
+        return {
+          assets,
+          bundleData:
+            data?.owner?.repository?.branch?.head?.bundleAnalysisReport?.bundle
+              ?.bundleData,
+          pageInfo:
+            data?.owner?.repository?.branch?.head?.bundleAnalysisReport?.bundle
+              ?.assetsPaginated?.pageInfo ?? null,
+        }
       }),
-    enabled: enabled,
+    getNextPageParam: (data) => {
+      return data?.pageInfo?.hasNextPage ? data?.pageInfo?.endCursor : undefined
+    },
+    enabled: opts?.enabled !== undefined ? opts.enabled : true,
     suspense: !!opts?.suspense,
   })
+
+  return {
+    data: {
+      assets: data?.pages.map((page) => page.assets).flat() ?? [],
+      bundleData: data?.pages?.[0]?.bundleData ?? null,
+    },
+    ...rest,
+  }
 }
